Optimize `scatter_search.rs`
Optimize `scatter_search.rs` by:
- Precomputing distances
- Only comparing new reference point with the newest
- Use `swap_remove` for O(1) candidate removal instead of O(n) retain

It uses more memory because of the `min_dists` vector but it is faster.

It should be possible to replace the vector of 1D arrays with a 2D ndarray for the reference set.  We could benefit from potential vectorized operations, but the dynamic aspects of the algorithm might become harder.