Add `max_time` parameter to `OQNLP.rs`
Adds a simple parameter `max_time`:
Max time for the stage 2 of the OQNLP algorithm.
If the time limit is reached, the algorithm will stop and return the solution set found so far.
The time limit is in seconds and it starts timing after the first local search.
It is optional and can be set to `None` to disable the time limit.

Also added better error handling and tests.