Add `progress_bar` feature
Add `progress_bar` feature that uses the `kdam` create to display a progress bar for the stage two of the algorithm. It is useful for large scale problems. I also added a test for this functionality.

Update README.md to show this change.

At some point I would also like to add a progress bar for the stage 1 part of the algorithm.