Use Euclidean Distance Squared
This commit changes Euclidean Distance to Euclidean Distance Squared.

We compare distances squared instead of distances, so we avoid calculating the square root which should be a performance improvement.

Add a test.

We have to be careful of numerical stability issues (underflow/overflow), maybe we should add some type of checks.