Add support for constraints in PyProblem
Introduces an optional constraints field to PyProblem, updates type hints and docstrings, and implements global constraint registry and dynamic constraint function generation for COBYLA support.

I am not sure this is the correct way to do this. COBYLA expects function pointers, and I have hardcoded a maximum of 1000 constraints in Python because of compilation times.

Modified the default local solver to be COBYLA, and related Rust code is updated to handle Python constraint functions and pass them to the solver. Also update pyo3 and removed deprecated functions.